airline-api/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
